Forms Update
-Change form name to "userForm"
-Move from to the components folder "form"
-Close Button Visibility
-Made The modal should close when the user clicks outside of it for better user experience
-The user form at the top of the modal; is properly positioning for improved layout
-Modal Title Design is fix
- Color fix now Ensure that all color codes used are consistent with the overall application colors
